Fix mysterious fatal bug via :aot; Rename to lein-otf (jit is ambiguous!)
Somehow, the released version of lein-jit (which I had been using)
completely and utterly failed when I tried to use it again recently. The
loader never made it into the jar as either a .clj or .class! Perhaps the
problem was specifying a :main that wasn't in the target project's own
source tree... but again, it *used to work*. Tried using different versions
of Clojure, checkout-dependencies, deps vs. dev-deps... nothing worked.

I've fixed this by adding the loader to :aot.